fix accidential state reuse due to mutable default dict parameter (#45)
Before this change, the NIR interpreter created a torch.fx.Graph with a `state` parameter, which had a dictionary
as a default value. Because Python evaluates default function parameters at definition time, the default
value would be reused across multiple calls when forward() is called without a state parameter.

The solution used here is the standard solution for this problem: Make the default value None,
and then set the value to the default dict inside the function body.

A side-effect of this change is that forward() can now be called with the state parameter
set to None, which eliminates the previously required conditional call to forward without
state parameter in the first time step.